added encoder and its mnist test
